SLIM-882 Fixes replace keys test
